% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content_webels.r
\name{get_content_webels}
\alias{get_content_webels}
\title{Get the content from all webelements in a list}
\usage{
get_content_webels(webels, contenttype = "text")
}
\arguments{
\item{webels}{A list with webelements}

\item{contenttype}{A character string to choose the content to extract: Defaults to "text", "url" would be the other option}
}
\value{
A vector containing the text/urls of all webelements
}
\description{
This function allows you to extract all text / urls  from a list of webelments.
}
\examples{
url <- "https://www.hyundai.ch/"
remDr <-  prepare_browser_function(url,"id","onetrust-reject-all-handler", containerport = 4445L)
webelements <- remDr$findElements(using = "class", "menu")
result <-  get_content_webels(webelements, )
}
