% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_href_webels.r
\name{get_href_webels}
\alias{get_href_webels}
\title{Get all urls from list of webelements}
\usage{
get_href_webels(webels)
}
\arguments{
\item{webels}{A list with webelements}
}
\value{
A vector containing the urls of all webelements
}
\description{
This function allows you to extract all href elements from a list of webelments.
}
\examples{
url <- "https://www.hyundai.ch/"
remDr <-  prepare_browser_function(url,"id","onetrust-reject-all-handler", containerport = 4445L)
webelements <- remDr$findElements(using = "class", "menu")
result <-  get_href_webels(webelements)
}
