% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text_webels.r
\name{get_text_webels}
\alias{get_text_webels}
\title{Get the text from all webelements in a list}
\usage{
get_text_webels(webels)
}
\arguments{
\item{webels}{A list with webelements}
}
\value{
A vector containing the text of all webelements
}
\description{
This function allows you to extract all text elements from a list of webelments.
}
\examples{
url <- "https://www.hyundai.ch/"
remDr <-  prepare_browser_function(url,"id","onetrust-reject-all-handler", containerport = 4445L)
webelements <- remDr$findElements(using = "class", "menu")
result <-  get_text_webels(webelements)
}
