% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_visible_webels.r
\name{get_visible_webels}
\alias{get_visible_webels}
\title{Get only the visible elements from a list of webelements}
\usage{
get_visible_webels(list_of_webel)
}
\arguments{
\item{webels}{A list with webelements}
}
\value{
A vector containing only the visible webelements
}
\description{
This function allows you to extract only the visible elements from a list of webelments.
}
\examples{
url <- "https://www.hyundai.ch/"
remDr <-  prepare_browser_function(url,"id","onetrust-reject-all-handler", containerport = 4445L)
remDr$navigate("https://www.hyundai.ch/de/model/i10/")
configbutton <- remDr$findElements(using = "id", "configBtn")
visbible_configbutton <- get_visible_webels(configbutton)
length(configbutton)
length(visbible_configbutton)
}
