% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_function.r
\name{insert_function}
\alias{insert_function}
\title{Opens a Database Connection and sends a SQL Statement to the database}
\usage{
insert_function(sqlstring)
}
\arguments{
\item{url}{A string which contains a SQL Statement}
}
\description{
This function allows you to send a SQL Statement to the database.
}
\examples{
sqlstring <- "SELECT * FROM ConfigCars"
insert_function(sqlstring)
}
