% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_browser_function.r
\name{prepare_browser_function}
\alias{prepare_browser_function}
\title{Open and Prepare Browser in Docker Container}
\usage{
prepare_browser_function(
  url,
  using_arg,
  search_arg,
  containerport,
  duration = 10000
)
}
\arguments{
\item{url}{A string which contains a url}

\item{using_arg}{A string which contains the Locator Scheme. As Example 'css selector', 'xpath',etc.}

\item{search_arg}{A string which contains the search term like a xpath}

\item{containerport}{A integer which identfies the Docker Container Port}

\item{duration}{A integer to specify the timeout duration}
}
\value{
A RemoteDriver Connection to the prepared Broswer
}
\description{
This function allows you open the browser with a specific url and sets the
Pageload/Implicit timeout. There is a option that cookie will be clicked.
}
\examples{
url <- "https://www.audi.ch/ch/web/de/neuwagen.html"
remDr <- prepare_browser_function(url,"id","onetrust-accept-btn-handler",containerport = 4445L)
}
