% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_sql_function.r
\name{row_sql_function}
\alias{row_sql_function}
\title{Writes a dataframe to the database}
\usage{
row_sql_function(df_raw, tablename)
}
\arguments{
\item{df_raw}{A dataframe or a matrix}

\item{tablename}{A string which contains the name of table in the database}
}
\description{
This function allows you to write the data of dataframe to the database.
}
\examples{
df <-  data.frame(a= "audi s6 bezin", b="80000 CHF")
row_sql_function(df, "ConfigCars")
}
