% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seperate_price_from_string.r
\name{seperate_price_from_string}
\alias{seperate_price_from_string}
\title{Seperate Price from/in a string}
\usage{
seperate_price_from_string(string_vector, price = TRUE)
}
\arguments{
\item{string_vector}{A string which contains a price with currency CHF}

\item{price}{A logical value to indicate if the price or the rest should be the result. Defaults to TRUE}
}
\value{
A string which containts the result
}
\description{
This function allows you open the browser with a specific url and sets the
}
\examples{
result <- seperate_price_from_string("Navigation CHF 1000")
}
