% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{test}
\alias{test}
\title{A Connection to the Database}
\usage{
test()
}
\value{
connectToDb_func() returns an S4 object that inherits from DBIConnection.
This object is used to communicate with the database engine.
}
\description{
This function allows you to connect to the MySql Database.
}
\examples{
libary(DBI)
con <- connectToDb_func()
sqlstring <- "SELECT * FROM ConfigCars"
df <- dbGetQuery(con, sqlstring)
}
