% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scroll_into_view.R
\name{scroll_into_view}
\alias{scroll_into_view}
\title{Scroll a webelemt into the view}
\usage{
scroll_into_view(rD, webel)
}
\arguments{
\item{rD}{A Remote Driver Connection to the browser in docker container}

\item{webel}{A Selenium Webelement}
}
\description{
This function allows you to scroll webelements into the view
}
