% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scroll_into_view.R
\name{scroll_into_view}
\alias{scroll_into_view}
\title{Scroll a webelemt into the view}
\usage{
scroll_into_view(rD, webel)
}
\arguments{
\item{rD}{A Remote Driver Connection to the browser in docker container}

\item{webel}{A Selenium Webelement}
}
\description{
This function allows you open the browser with a specific url and sets the
Pageload/Implicit timeout. There is a option that cookie will be clicked.
}
\examples{
url <- "https://www.hyundai.ch/"
remDr <-  prepare_browser_function(url,"id","onetrust-reject-all-handler", containerport = 4445L)
webelement <- remDr$findElement(using = "class", "menu")
scroll_into_view(remDr, webelement)
}
